render objects

+++ src/renderdata/include/renderdata/zwvertex.h
@@ -68,7 +68,7 @@ const std::vector<ZwVertex> zwVertices =
 	ZwVertex(glm::vec3(-0.5f, 0.5f, -0.5f), glm::vec3(1.0f, 1.0f, 1.0f), glm::vec2(1.0f, 1.0f)),
 };
 
-const std::vector<uint16_t> zwIndices = 
+const std::vector<uint32_t> zwIndices = 
 {
 	0, 1, 2, 2, 3, 0,
 	4, 5, 6, 6, 7, 4
